requirements.txt
Django==5.0.7
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
django-filter==24.2
psycopg2-binary==2.9.9
python-dotenv==1.0.1
